Saturday night commit
You would not have believed me a few weeks ago if I told you that I would work on the Software Engineering project on a Saturday night. But as I started getting deeper into this project, I found myself almost addicted to it, sometimes even forgetting to interact with the external world just to add a few lines to the DD. Or maybe I am just quarantined and this is the most entertaining activity of this Saturday night.

P.S. the DD is almost ended